pubg-scrapper/requirements.txt
selenium==4.20.0
webdriver-manager==4.0.1
kafka-python==2.0.2
requests==2.31.0
